Create AB Testing for Shoefly.com - Codecademy Data Analyst Path Project
Project as part of the Data Analyst path on Codecademy.com. Description:

'Our favorite online shoe store, ShoeFly.com is performing an A/B Test. They have two different versions of an ad, which they have placed in emails, as well as in banner ads on Facebook, Twitter, and Google. They want to know how the two ads are performing on each of the different platforms on each day of the week. Help them analyze the data using aggregate measures.'